refactor: extract shared SQL-fixture-queue + STABLE_EMBEDDING test helpers
Closes #421. Pulls the SQL-fixture-queue test harness (used by entity-resolution
tests to mock transactional() against substring-keyed fixtures) and the
STABLE_EMBEDDING constant into two shared modules under src/test/, and migrates
all five callsites listed in the issue to import from them. Removes ~330 lines
of duplicated harness code with no behavior change.

The vi.mock("@/db/pool", factory) call must remain inline per consumer file
(Vitest hoists it above all imports), but the factory body now dynamic-imports
the shared helpers — top-level imports referenced from the factory closure
fail with "cannot access ... before initialization" since the factory runs
before the test file's import phase completes. Documented in the harness
header.
